==> Project_23/4_VerityWorkflow.Rmd <==
---
title: "Clean Verity workflow"
output: html_document
date: "2023-07-18"
editor_options: 
  markdown: 
    wrap: 72
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Package management

```{r}

update.packages(ask = FALSE)

install.packages("BiocManager")

BiocManager::install(c(
  "org.Mm.eg.db",
  "AnnotationDbi",
  "clusterProfiler",
  "tidyverse",
  'knitr',
  "DropletUtils",
  "Seurat",
  "SoupX",
  "scCustomize",
  "glmGamPoi",
  "DESeq2",
  "Polychrome",
  "ReactomePA"
  ))


remotes::install_github('chris-mcginnis-ucsf/DoubletFinder')

remotes::install_github("satijalab/seurat", "seurat5")

devtools::install_github("junjunlab/GseaVis")

```

```{r}
library(org.Mm.eg.db)
library(clusterProfiler)
library(DropletUtils)
library(Seurat)
library(tidyverse)
library(dplyr)
library(stringr)
library(knitr)
library(DoubletFinder)
library(SoupX)
library(scCustomize)
library(DESeq2)
library(Polychrome)
library(GseaVis)
library(ReactomePA)
library(cowplot)
library(biomaRt)
```

Make folder for plots and set seed

```{r}
#save folder path for plot saving
plotting_path_Ver<-paste0("./Verity/Images/")

#set seed for reproducibility
set.seed(42)

```

#Verity workflow

This workflow loads samples for the control and prion-treated data from
Mabbott group's hippocampal prion work. The samples (not including the
ageing samples sample9 and 11) are QC assessed, filtered and plotted
with clustering by gene expression profiles.

Adapted from
[here](https://bioconductor.org/packages/release/bioc/vignettes/DropletUtils/inst/doc/DropletUtils.html#reading-in-10x-genomics-data)
and
[here](https://www.singlecellcourse.org/scrna-seq-analysis-with-bioconductor.html).

```{r}
#load Nick Verity's data 

#Comment out the samples which aren't needed for the comparison of interest

# List of folder names
folder_names <- c(
  "sample1", "sample2",
  "sample5", "sample6",
  "sample9", "sample11"
  )

# Loop through each folder
for (folder_name in folder_names) {
  folder_path=paste0("/exports/cmvm/eddie/eb/groups/mabbott_grp/NVerity/snRNAseq/expdata/A231222/combined/",folder_name, "/DGE_unfiltered")
  
  #Load the matrix file
  current_parse <- ReadParseBio(folder_path)
  
  #see if there are any empty gene names
  table(rownames(current_parse) == "") 
  
  #Load the metadata
  current_metadata<-read.csv(paste0(folder_path,"/cell_metadata.csv"))

  #Create a seurat object
  current_parse_interpreted <- CreateSeuratObject(current_parse, meta.data = current_metadata, names.delim = "")
  
  #Save the cell names as metadat and set identities to the sample
  current_parse_interpreted@meta.data$cell_name<-rownames(current_parse_interpreted@meta.data)
  current_parse_interpreted@meta.data$sample <- factor(rep(folder_name, nrow(current_parse_interpreted@meta.data)))
  
  Idents(current_parse_interpreted) <- current_parse_interpreted@meta.data$sample


  # Assign the object to a variable with the folder name as the variable name
  assign((paste0(folder_name, "_obj")),current_parse_interpreted)

}


#merge the samples into one object
Verity_total_clean <- merge(sample1_obj, y=c(sample2_obj, sample5_obj, sample6_obj,sample9_obj, sample11_obj), add.cell.ids=c("sample1", "sample2", "sample5", "sample6", "sample9", "sample11"), project="Verity")

#cleaning up memory
rm("sample1_obj","sample2_obj", "sample5_obj", "sample6_obj", "sample9_obj", "sample11_obj",current_metadata, current_parse_interpreted, current_parse)
```

# Tailor metadata

```{r}
metadata_to_remove<-c("bc_wells","species","gene_count","tscp_count", "mread_count","bc1_well","bc2_well","bc3_well","bc1_wind","bc2_wind", "bc3_wind")

for (var in metadata_to_remove){
  Verity_total_clean[[var]]<-NULL
}

```


## treatment annotation

adapted from <https://youtu.be/p49seH2_i8Y?t=312>

```{r}
sampleLabs<-names(Verity_total_clean@active.ident)

#Assigning names to the samples to keep identity
#From Chat GPT
treat_detect <- case_when(
  str_detect(sampleLabs, "sample11") ~ "AGEING", #assign this treatment condition first to avoid mis-labelling with sample1's label via str_detect
  str_detect(sampleLabs, "sample1") ~ "NBH",
  str_detect(sampleLabs, "sample2") ~ "NBH",
  str_detect(sampleLabs, "sample5") ~ "ME7",
  str_detect(sampleLabs, "sample6") ~ "ME7",
  str_detect(sampleLabs, "sample9") ~ "AGEING",
  TRUE ~ NA_character_
)

#Save as metadata
Verity_total_clean@meta.data$treatment<-treat_detect
rm(sampleLabs, treat_detect)
```

```{r}
#plotting colours dataframe
colour_df_Verity<-data.frame(folder_names)
colour_df_Verity$colours<-c('#59b47b',
'#58a92f',
'#ae87ff',
'#ee623f',
'#b6923c',
'#f5496b')

colour_df_Verity_prion<-subset(colour_df_Verity, subset = folder_names != 'sample11' &folder_names != 'sample9')
colour_df_Verity_ageing<-subset(colour_df_Verity, subset = folder_names == 'sample11' | folder_names == 'sample9')
```


## Identify mitochondrial and ribosomal populations

```{r}
#ID mitochondrial dominant cells 
Verity_total_clean[["percent.mt"]]<- PercentageFeatureSet(Verity_total_clean, pattern = "^mt")
Verity_total_clean[["percent.rp"]]<- PercentageFeatureSet(Verity_total_clean, pattern = "^Rp")
Verity_total_clean[["percent.hb"]]<-PercentageFeatureSet(Verity_total_clean, "^Hb[^(p)]")

#Generate genes per umi counts, from https://github.com/hbctraining/scRNA-seq/blob/master/lessons/04_SC_quality_control.md
Verity_total_clean$log10GenesPerUMI <- log10(Verity_total_clean$nFeature_RNA) / log10(Verity_total_clean$nCount_RNA)

Verity_total_clean$mitoRatio <- Verity_total_clean@meta.data$percent.mt / 100


View(Verity_total_clean@meta.data)
```

##Initial data assessment

```{r}
p1<-VlnPlot(Verity_total_clean, features=c("nFeature_RNA","nCount_RNA","percent.mt", "percent.rp","percent.hb"), ncol=5, pt.size = 0.0)+scale_fill_manual(values = colour_df_Verity$colours)+plot_annotation(title = "Verity dataset EDA")
p1
ggsave("Verity_EDA.png",plot=p1,device='png', path=plotting_path_Ver,width = 40, height = 15, units = 'cm' )

# Plot each of these respectively, showing trends in cell health

#show mitochondrial proportion in reads
p2<-FeatureScatter(Verity_total_clean, feature1 = "nCount_RNA", feature2 = "percent.mt", pt.size = 0)+scale_fill_manual(values = colour_df_Verity$colours)+plot_annotation(title = "Verity mitochondrial genes detected")
p2
ggsave("Verity_UMI_MT.png",plot=p2,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm' )


#show ribosomal proportion in reads
p3<-FeatureScatter(Verity_total_clean, feature1 = "nCount_RNA", feature2 = "percent.rp")+scale_fill_manual(values = colour_df_Verity$colours)+plot_annotation(title = "Verity_UMIs_per_cell and ribosomal_protein_genes_detected")
p3
ggsave("Verity_UMI_RP.png",plot=p3,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm' )


#Ribosomal vs mitochondrial 
p4<-FeatureScatter(Verity_total_clean, feature1 = "percent.rp", feature2 = "percent.mt")+scale_fill_manual(values = colour_df_Verity$colours)+plot_annotation(title = "Verity_ribosomal_protein_genes_detected_vsmitochondrial_genes_detected")
p4
ggsave("Verity_RP_MT.png",plot=p4,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm' )

#Show overall reads and gene counts for unfiltered data.
p5<-FeatureScatter(Verity_total_clean, feature1 = "nCount_RNA", feature2="nFeature_RNA",raster=F, pt.size = 0.5)+ geom_smooth(method="lm")+scale_fill_manual(values = colour_df_Verity$colours)+plot_annotation(title = "Verity_UMIs_per_cell_vs_number_of_genes")
p5
ggsave("Verity_UMI_nGene.png",plot=p5,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm' )

```

### Save the whole data

```{r}
#Save the data object
save(Verity_total_clean, file="./Verity/Data/Verity_total_clean.Rdata")
gc()
```

#QC

##Exploration

### Create metadata variable to edit it without affecting core data

```{r}
# Rename columns from https://github.com/hbctraining/scRNA-seq/blob/master/lessons/04_SC_quality_control.md

#Create metadata variable to edit it without affecting core data
metadata_test_Verity<-Verity_total_clean@meta.data

#Saving cell barcodes as metainfo
metadata_test_Verity$cells<-rownames(metadata_test_Verity)

#Setting minimum feature count to filter cells with no genes UMIs vs. genes detected, as these will be discarded anyway
metadata_test_Verity <- subset(Verity_total_clean, subset= nFeature_RNA > 20)@meta.data

metadata_test_Verity <- metadata_test_Verity %>%
    dplyr::rename(nUMI = nCount_RNA,
                  nGene = nFeature_RNA)



```

### Plots

```{r}
# Visualize the number of cell counts per sample with no filtering at all
p6_ver<-metadata_test_Verity %>% 
  	ggplot2::ggplot(aes(x=sample, fill=sample)) + 
  	geom_bar() +
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation(title = "Verity_Cell_counts_per_sample")+
    guides(fill = 'none')

p6_ver

# Visualize the number UMIs/transcripts per cell
p7_ver<-metadata_test_Verity %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=nUMI, fill= sample)) + 
  	geom_density(alpha = 0.2) + 
  	scale_x_log10() + 
  	theme_classic() +
  	ylab("Cell density") +
  	geom_vline(xintercept = 500)+
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation(title = "Verity/UMI_transcripts_per_cell")
    #guides(fill = 'none')
p7_ver

# Visualize the distribution of genes detected per cell via histogram
p8_ver<-metadata_test_Verity %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=nGene, fill= sample)) + 
  	geom_density(alpha = 0.2) + 
  	theme_classic() +
  	scale_x_log10() + 
  	geom_vline(xintercept = 300)+
    scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation(title = "Verity distribution of genes per cell")
    #guides(fill = 'none')

p8_ver

# Visualize the distribution of genes detected per cell via boxplot
p9_ver<-metadata_test_Verity %>% 
  	ggplot(aes(x=sample, y=log10(nGene), fill=sample)) + 
  	geom_boxplot() + 
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
    scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation(title = "Verity distribution of genes per cell")+
    guides(fill = 'none')
p9_ver    

# Visualize the correlation between genes detected and number of UMIs and determine whether strong presence of cells with low numbers of genes/UMIs
p10_ver<-metadata_test_Verity %>% 
  	ggplot(aes(x=nUMI, y=nGene, color=mitoRatio)) + 
  	geom_point() + 
	scale_colour_gradient(low = "gray90", high = "black") +
  	scale_x_log10() +
  	scale_y_log10() +
  	theme_classic() +
  	geom_vline(xintercept = 500) +
  	geom_hline(yintercept = 250) +
  	facet_wrap(~sample) +
    scale_fill_manual(values = colour_df_Verity$colours)+
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
    plot_annotation(title = "Verity UMIs vs. genes detected. Highlighting mitochondrial ratio in genes")
  
p10_ver


# Visualize the distribution of mitochondrial gene expression detected per cell
p11_ver<-metadata_test_Verity %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=mitoRatio, fill=sample)) + 
  	geom_density(alpha = 0.2) + 
  	scale_x_log10() + 
  	theme_classic() +
  	geom_vline(xintercept = 0.2)+
    scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation(title = "Verity mitochondrial ratio per sample")
  
p11_ver

ggsave("Verity_PreF_Counts_sample.png",plot=p6_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm' )
ggsave("Verity_PreF_UMI_cell.png",plot=p7_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PreF_gene_cell.png",plot=p8_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PreF_gene_cell_boxplot.png",plot=p9_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PreF_MT_cell_UMI.png",plot=p10_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PreF_MT_cell.png",plot=p11_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

```

## Filtering metadata

```{r}
metadata_filtered<- metadata_test_Verity %>%
    filter(nUMI > 500, nGene >1100, nUMI <35000, percent.mt <20, percent.hb < 20) 
```

##Repeat QC plotting

```{r}
# Visualize the number of cell counts per sample with no filtering at all
p12_ver<-metadata_filtered %>% 
  	ggplot2::ggplot(aes(x=sample, fill=sample)) + 
  	geom_bar() +
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity_Cell_counts_per_sample", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")+
    guides(fill = 'none')
    

# Visualize the number UMIs/transcripts per cell
p13_ver<-metadata_filtered %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=nUMI, fill= sample)) + 
  	geom_density(alpha = 0.2) + 
  	scale_x_log10() + 
  	theme_classic() +
  	ylab("Cell density") +
  	geom_vline(xintercept = 500)+
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity_UMI_transcripts_per_cell", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")
    #guides(fill = 'none')


# Visualize the distribution of genes detected per cell via histogram
p14_ver<-metadata_filtered %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=nGene, fill= sample)) + 
  	geom_density(alpha = 0.2) + 
  	theme_classic() +
  	scale_x_log10() + 
  	geom_vline(xintercept = 300)+
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity distribution of genes per cell", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")
    #guides(fill = 'none')
    


# Visualize the distribution of genes detected per cell via boxplot
p15_ver<-metadata_filtered %>% 
  	ggplot(aes(x=sample, y=log10(nGene), fill=sample)) + 
  	geom_boxplot() + 
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity distribution of genes per cell", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")+
    guides(fill = 'none')
    

# Visualize the correlation between genes detected and number of UMIs and determine whether strong presence of cells with low numbers of genes/UMIs
p16_ver<-metadata_filtered %>% 
  	ggplot(aes(x=nUMI, y=nGene, color=mitoRatio)) + 
  	geom_point() + 
	scale_colour_gradient(low = "gray90", high = "black") +
  	scale_x_log10() +
  	scale_y_log10() +
  	theme_classic() +
  	geom_vline(xintercept = 500) +
  	geom_hline(yintercept = 250) +
  	facet_wrap(~sample) +
  	#scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity UMIs vs. genes detected. Highlighting mitochondrial ratio in genes", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")
    #guides(fill = 'none')
    

# Visualize the distribution of mitochondrial gene expression detected per cell
p17_ver<-metadata_filtered %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=mitoRatio, fill=sample)) + 
  	geom_density(alpha = 0.2) + 
  	scale_x_log10() + 
  	theme_classic() +
  	geom_vline(xintercept = 0.2)+
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity distribution of mitochondrial gene expression per cell", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")
    #guides(fill = 'none')
    
    
ggsave("Verity_PostF_Counts_sample.png",plot=p12_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PostF_UMI_cell.png",plot=p13_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PostF_nGene_hist_sample.png",plot=p14_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PostF_nGene_sample_boxplot.png",plot=p15_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PostF_MT_cell_UMI.png",plot=p16_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_PostF_MT_cell.png",plot=p17_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

p12_ver;p13_ver;p14_ver;p15_ver;p16_ver;p17_ver

```

## Assign filters to data

```{r}
#Assign the unfiltered metadata to Verity, updating the column names
Verity_total_clean@meta.data<-metadata_test_Verity 

Verity_filtered_clean<- subset(Verity_total_clean,subset= 
                              percent.mt <20 &
                              nUMI > 500 &
                              nUMI < 35000 &
                              nGene >1100 & 
                              percent.hb < 20
                            )


# Gene-level filtering
# Keeping genes which appear in more than 3 cells for statistical power.

# Output a logical vector for every gene on whether the more than zero counts per cell
# Extract counts
counts <- GetAssayData(object = Verity_filtered_clean, slot = "counts")

# Output a logical vector for every gene on whether the more than zero counts per cell
nonzero <- counts > 0

# Sums all TRUE values and returns TRUE if more than 3 TRUE values per gene
keep_genes <- Matrix::rowSums(nonzero) >= 3

# Only keeping those genes expressed in more than 10 cells
filtered_counts <- counts[keep_genes, ]

# Reassign to filtered Seurat object
Verity_filtered_clean <- CreateSeuratObject(filtered_counts, meta.data = Verity_filtered_clean@meta.data)

p18_ver<-Verity_filtered_clean@meta.data %>% 
  	ggplot2::ggplot(aes(x=sample, fill=sample)) + 
  	geom_bar() +
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity_Cell_counts_per_sample", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")+
    guides(fill = 'none')
    
ggsave("Verity_Final_sample_counts.png",plot=p18_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')


p19_ver<-Verity_filtered_clean@meta.data %>% 
  	ggplot(aes(
  	  #color=sample, 
  	  x=nGene, fill= sample)) + 
  	geom_density(alpha = 0.2) + 
  	theme_classic() +
  	scale_x_log10() + 
  	geom_vline(xintercept = 300)+
  	scale_fill_manual(values = colour_df_Verity$colours)+
    plot_annotation("Verity distribution of genes per cell", subtitle="nGene>1100, 500<nUMI<35000, % mito< 20, % hb< 20")
    
ggsave("Verity_Final_gene_sample.png",plot=p19_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

p18_ver;p19_ver

rm(metadata_test_Verity, metadata_filtered,filtered_counts, nonzero,counts, keep_genes,folder_name,folder_names,folder_path,metadata_to_remove,var)
gc()
```

### Saving object and tidying up

```{r}
# Create .Rdata object to load at any time
save(Verity_filtered_clean, file="./Verity/Data/Verity_filtered_clean.Rdata")
```

# DoubletFinder

##Splitting the data by treatment

```{r}
Verity_Prion_comp<-subset(Verity_filtered_clean, subset=treatment==c('NBH','ME7'))
Verity_Ageing_comp<-subset(Verity_filtered_clean, subset=treatment==c('NBH','AGEING'))

Verity_comparisons<-list(Verity_Prion_comp,Verity_Ageing_comp)

rm(Verity_Prion_comp,Verity_Ageing_comp, Verity_filtered_clean, Verity_total_clean)
gc()
```

## Doublet finder for treatment comparisons separately

```{r}

#load in the filtered Verity object
#load(paste0(plotting_path,"Verity_filtered_clean.Rdata")

#adapted from https://youtu.be/p49seH2_i8Y , https://github.com/kpatel427/YouTubeTutorials/blob/5b3c795cecafb22f0e992da069eae2efd6b5cb95/singleCell_doublets.R , https://rpubs.com/kenneditodd/doublet_finder_example 

#DoubletFinder, keep singlets
Verity_comparisons <- lapply(Verity_comparisons, function(comp_pair) {
  
  #Normalise+Scale with SCTransform?
  comp_pair <- SCTransform(comp_pair)
  comp_pair <- RunPCA(comp_pair)
  
  # Find significant PCs
  stdv <- comp_pair[["pca"]]@stdev
  sum.stdv <- sum(comp_pair[["pca"]]@stdev)
  percent.stdv <- (stdv / sum.stdv) * 100
  cumulative <- cumsum(percent.stdv)
  co1 <- which(cumulative > 90 & percent.stdv < 5)[1]
  co2 <- sort(which((percent.stdv[1:length(percent.stdv) - 1] -
                       percent.stdv[2:length(percent.stdv)]) > 0.1),
              decreasing = TRUE)[1] + 1
  min.pc <- min(co1, co2)
  
  # Finish pre-processing
  comp_pair <- RunUMAP(comp_pair, dims = 1:min.pc, seed.use=42)
  comp_pair <- FindNeighbors(object = comp_pair, dims = 1:min.pc)              
  comp_pair <- FindClusters(object = comp_pair, resolution = 0.1)
  
  # pK identification (no ground-truth)
  sweep.list <- paramSweep_v3(comp_pair, PCs = 1:min.pc, sct = TRUE)
  sweep.stats <- summarizeSweep(sweep.list)
  bcmvn <- find.pK(sweep.stats)
  
  # Optimal pK is the max of the bomodality coefficient (BCmvn) distribution
  bcmvn.max <- bcmvn[which.max(bcmvn$BCmetric),]
  optimal.pk <- as.numeric(bcmvn.max$pK)

  # Homotypic doublet proportion estimate
  annotations <- comp_pair@meta.data$seurat_clusters
  homotypic.prop <- modelHomotypic(annotations) 
  nExp.poi <- round(0.03 * nrow(comp_pair@meta.data)) #expected doublets based on values from PMC7424855
  nExp.poi.adj <- round(nExp.poi * (1 - homotypic.prop))
  
  # Run DoubletFinder
  comp_pair <- doubletFinder_v3(seu = comp_pair, 
                             PCs = 1:min.pc, 
                             pK = optimal.pk,
                             nExp = nExp.poi.adj, 
                             sct = TRUE, 
                             pN=0.25)
  
  return(comp_pair)
})
```

#Show overall QC stats after filtering

```{r}
load(paste0(plotting_path_Ver,"Verity_comparisons.Rdata"))
Verityqc_comp1<-Verity_comparisons[[1]]
Verityqc_comp2<-Verity_comparisons[[2]]

Verity_Qc_comp<-merge(x=Verityqc_comp1, y=Verityqc_comp2,project="Verity")

Idents(Verityqc_comp1)<-Verityqc_comp1$sample

Verity_comparisons_filtered_QC<-VlnPlot(Verityqc_comp1, features=c("nFeature_RNA","nCount_RNA","percent.mt", "percent.rp","percent.hb"), ncol=5, pt.size = 0.0)&scale_fill_manual(values = colour_df_Verity_prion$colours)&plot_annotation(title = "Verity dataset after all filtering")



Verity_comparisons_filtered_barplot<-Verityqc_comp1@meta.data%>%
    ggplot(aes(x=factor(sample,level=c("sample1","sample2", "sample5", "sample6")), fill=sample)) + 
  	geom_bar() +
  	theme_classic() +
  	theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  	theme(plot.title = element_text(hjust=0.5, face="bold")) +
    scale_fill_manual(values = colour_df_Verity_prion$colours)+
    scale_y_continuous(breaks=seq(0, 30000, 3000))+
    guides(fill = 'none')
Verity_comparisons_filtered_barplot

p20_ver<-Verity_comparisons_filtered_QC+
  Verity_comparisons_filtered_barplot+
  ggtitle("Cells per sample")+
  plot_layout(ncol = 3)&
  plot_annotation(title = "Verity dataset after all filtering")&
  xlab(NULL)

p20_ver

ggsave("Verity_comparisons_filtered_QC.png",plot=p20_ver,device='png', path=plotting_path_Ver,width = 30, height = 15, units = 'cm' )

rm(Verityqc_comp1,Verityqc_comp2,Verity_Qc_comp)
```


# Split the data, rename doubletfinder metadata and Show Doublets for each comparison group witihn Verity data. Retain singlets

```{r}

#Rename the columns
Verity_comparisons <- lapply(Verity_comparisons, function(comp_pair) {
  DefaultAssay(comp_pair)<-'RNA'
  doublet_col_index <- which(grepl("DF.classifications_", colnames(comp_pair@meta.data)))
  comp_pair$doublet_finder <- comp_pair@meta.data[doublet_col_index]
  comp_pair@meta.data[doublet_col_index]<-NULL

  #Show how many doublets there were
  table(comp_pair$doublet_finder)
  
  return(comp_pair)
})

#Make suer the samples are properly assigned to the Prion treatment group
Verity_Prion_comp<-subset(Verity_comparisons[[1]], subset=sample !="sample11")

#Split the data
Verity_Prion_comp<-Verity_comparisons[[1]]
Verity_Ageing_comp<-Verity_comparisons[[2]]

#Plot how many doublets there were
p21_ver<-VlnPlot(Verity_Prion_comp, features = "nFeature_RNA", group.by = 'doublet_finder', pt.size = 0.0)+scale_fill_manual(values = colour_df_Verity_prion$colours)+plot_annotation("Verity Control and Prion doublets")
p21_ver
ggsave("Verity_Prion_comp_doublets.png",plot=p21_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

p22_ver<-VlnPlot(Verity_Ageing_comp, features = "nFeature_RNA", group.by = 'doublet_finder', pt.size = 0.0)+scale_fill_manual(values = colour_df_Verity_ageing$colours)+plot_annotation("Verity Control and Ageing doublets")
p22_ver
ggsave("Verity_Ageing_comp_doublets.png",plot=p22_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')


#Save the data with Doublets retained too
save(Verity_comparisons, file="./Verity/Data/Verity_comparisons.Rdata")

#subset data by singlet status and save list of singlets to continue with
Verity_Prion_comp <- subset(Verity_Prion_comp, subset= doublet_finder=='Singlet')
Verity_Ageing_comp <- subset(Verity_Ageing_comp, subset= doublet_finder=='Singlet')


gc()
```

# Prion work

```{r}
#active assay RNA
DefaultAssay(Verity_Prion_comp)<-"RNA"

#DimReduc again
#https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/06_SC_SCT_normalization.md
# Normalize the counts
Verity_Prion_comp <- NormalizeData(Verity_Prion_comp)
# Identify the most variable genes with defaults
Verity_Prion_comp <- FindVariableFeatures(Verity_Prion_comp, verbose = T)
		     
# Scale the counts
Verity_Prion_comp <- ScaleData(Verity_Prion_comp)
# Perform PCA
Verity_Prion_comp <- RunPCA(Verity_Prion_comp)
#Inspect the variability accounted for by the PCS
#It appears that 23 PCs are responsible, so using 30 from now on
ElbowPlot(Verity_Prion_comp, ndims=30)+ggtitle("Elbowplot of Verity data", subtitle = "NBH and ME7")
DimHeatmap(Verity_Prion_comp, dims=1:9, cells=500, balanced=T)
#Finish dimensinoality reduction steps
Verity_Prion_comp <- RunUMAP(Verity_Prion_comp, reduction = "pca", dims = 1:30, seed.use=42)
Verity_Prion_comp <- FindNeighbors(Verity_Prion_comp, reduction = "pca", dims = 1:30)
Verity_Prion_comp <- FindClusters(Verity_Prion_comp, resolution = c(0.2,0.3,0.5,0.7,1))
#Save the object

save(Verity_Prion_comp,file="./Verity/Data/Verity_Prion_comp.Rdata")

# Visualise the clustering resolutions for cluster number

p23_ver<-DimPlot(Verity_Prion_comp , reduction = "umap", group.by = "treatment", raster=FALSE)+ggtitle(NULL)+ plot_annotation("Verity Hippocampal cells labelled by treatment")
p23_ver
ggsave("Verity_all_prion_treatment.png",plot=p23_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

#cycle/change resolution number in group.by to see difference in cluster boundaries
p24_ver<-DimPlot(Verity_Prion_comp, group.by = 'RNA_snn_res.0.2', reduction = 'umap', label = TRUE, repel = TRUE, raster=F)+ggtitle(NULL)+plot_annotation("Verity control and prion cells labelled by cluster numbers", subtitle = "Cluster resolution 0.2")
p24_ver
ggsave("Verity_all_prion_res02.png",plot=p24_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

#adapted from https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/08_SC_clustering_quality_control.md#segregation-of-clusters-by-various-sources-of-uninteresting-variation

# Determine metrics to plot present in Ver_singlets_merged_clean@meta.data
metrics <-  c("nUMI", "nGene", "mitoRatio")

#Explore any clustering mishaps by looking for localisation of any of these traits in a given cluster
p25_ver<-FeaturePlot(Verity_Prion_comp, 
            reduction = "umap", 
            features = metrics,
            pt.size = 0.4, 
            order = TRUE,
            label = TRUE)+plot_annotation(title="Verity control and prion cells, labelled by nUMI, nGene & mitochondrial ratio",tag_levels = 'A')
p25_ver+plot_layout(nrow=3, byrow=FALSE)
ggsave("Verity_all_prion_metrics_UMAP.png",plot=p25_ver,device='png', path=plotting_path_Ver,width = 40, height = 40, units = 'cm')


## Save the singlets with a cluster resolution?

 #set the idents to cluster resolution 0.2
Verity_Prion_comp@active.ident<-Verity_Prion_comp@meta.data$RNA_snn_res.0.2
levels(Verity_Prion_comp)

rm(metrics)
```

```{r}

DefaultAssay(Verity_Prion_comp)<-"RNA"

#Check for a gene of interest in the data's geneset
Gene_Names<-Verity_Prion_comp@assays[["RNA"]]@counts@Dimnames[[1]]
filtered_gene_names<-Gene_Names[!grepl("^\\d", Gene_Names)]
filtered_gene_names[grep("^Slc1a2", filtered_gene_names)]

#try to identify astrocytes  with genes
astrocyte_genes<-c("Aldh1l1","Gfap","Slc1a2", "Slc1a3", "Glul", "S100b")

Verity_astrocyte_markers_plot<-FeaturePlot(Verity_Prion_comp,features=astrocyte_genes, raster=F,order=T, label = T)&plot_annotation(title = "Verity prion data, labelled with astrocyte gene markers from ABCAM", subtitle="Clustering resolution 0.2, 30 PCs used")&scale_colour_gradient(low = "lightgrey", high = "blue",limits = c(0, 6))


#From ChatGPT
# Create the plot without the legend
plot_without_legend <- Verity_astrocyte_markers_plot + theme(legend.position = "none")
# Get the legend
legend <- get_legend(Verity_astrocyte_markers_plot)
# Combine the plot and legend using patchwork
p26_ver<- plot_without_legend + plot_layout(guides = "collect",ncol=3)
  #FIXME plot_annotation("Verity prion cells with astrocyte markers")
p26_ver
ggsave("Verity_astrocyte_markers_plot.png",plot=p26_ver,device='png', path=plotting_path_Ver,width = 40, height = 30, units = 'cm' )


#As a reminder...
DimPlot(Verity_Prion_comp, group.by = 'RNA_snn_res.0.2', reduction = 'umap', label = TRUE, repel = TRUE, raster=F)+ggtitle("Integrated Verity data, labelled by cluster numbers", subtitle="RPCA,Clustering resolution 0.3, 20 PCs used")

#Now compare to the plots of astrocyte genes
VlnPlot(Verity_Prion_comp, features = astrocyte_genes, group.by = 'RNA_snn_res.0.2', pt.size = 0.0)

#16 clusters
```

## (SingleR annotation)

Could go here, could be followed by bubbleplot
<https://rdrr.io/github/milescsmith/seuratBubblePlot/man/bubbleplot.html>

Alternatively, load information from the spreadsheet at
Analysis/ABCAM_markers.xlsx then bubble plot

##Investigate astrocytes, ensure identification

```{r}
# Convert cluster names to character
Verity_Prion_comp@active.ident <- as.factor(Verity_Prion_comp$RNA_snn_res.0.2)

#Subset out astrocytes
Verity_Prion_comp_astrocytes<-subset(Verity_Prion_comp, idents=c('6','11'))

save(Verity_Prion_comp,file="./Verity/Data/Verity_Prion_comp.Rdata")
Verity_Prion_comp_astrocytes
```

## DEG work

Prepare the biomart info

```{r}
#Adapted from http://girke.bioinformatics.ucr.edu/CSHL_RNAseq/mydoc/mydoc_systemPipeRNAseq_07/
#Get annotation info

m <- useMart("ensembl", dataset="mmusculus_gene_ensembl")
```

```{r}


# Run the standard workflow for visualization and clustering
Verity_Prion_comp_astrocytes <- ScaleData(Verity_Prion_comp_astrocytes, verbose = FALSE)
Verity_Prion_comp_astrocytes <- RunPCA(Verity_Prion_comp_astrocytes, npcs = 30, verbose = FALSE)
ElbowPlot(Verity_Prion_comp_astrocytes, ndims=30)+ggtitle("Elbowplot of subset Verity astrocytes", subtitle = "RunPCA npcs=30")

#15 PCs sems ok, so use 20 from now

#20 PCs seem to cover the variability
Verity_Prion_comp_astrocytes <- RunUMAP(Verity_Prion_comp_astrocytes, reduction = "pca", dims = 1:20, seed.use=42)
Verity_Prion_comp_astrocytes <- FindNeighbors(Verity_Prion_comp_astrocytes, reduction = "pca", dims = 1:20)
Verity_Prion_comp_astrocytes <- FindClusters(Verity_Prion_comp_astrocytes, resolution = c(0.2,0.3,0.5,0.7,1))

#save for reloading
save(Verity_Prion_comp_astrocytes, file="./Verity/Data/Verity_Prion_comp_astrocytes.Rdata")

p27_ver<-DimPlot(Verity_Prion_comp_astrocytes, reduction = 'umap',  raster=F, group.by='treatment')+ggtitle(NULL)+plot_annotation("Verity data astrocytes", subtitle="Labelled by treatment")
p27_ver
ggsave("Verity_prion_astrocyte_treatment_plot.png",plot=p27_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

#Show the clustering of subset astrocytes
p28_ver<-DimPlot(Verity_Prion_comp_astrocytes, group.by = 'RNA_snn_res.0.3', reduction = 'umap', repel = TRUE, raster=F)+ggtitle(NULL)+plot_annotation("Verity data astrocytes",subtitle=" Cluster resolution 0.3")
p28_ver
ggsave("Verity_prion_astrocyte_clusters.png",plot=p28_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')

#Show sample disctribution in clusters of astrocytes
DimPlot(Verity_Prion_comp_astrocytes, group.by = 'sample', reduction = 'umap', repel = TRUE, raster=F)

#Show astrocyte markers in astrocyte subset
Verity_reactive_astrocyte_plot<-FeaturePlot(Verity_Prion_comp_astrocytes,features=c('Gfap','C4b','Cd44', 'Osmr','Trf','Stat3','Aqp4', 'Nrg1', 'Gria1'), raster=F)&plot_annotation(title="Verity data astrocytes", subtitle="highlighting reactive astrocyte genes", tag_levels = 'A')&scale_colour_gradient(low = "lightgrey", high = "blue",limits = c(0,5))

plot_without_legend <- Verity_reactive_astrocyte_plot + theme(legend.position = "none")
# Get the legend
legend <- get_legend(Verity_reactive_astrocyte_plot)
# Combine the plot and legend using patchwork
p29_ver<- plot_without_legend + plot_layout(guides = "collect",ncol=3)
  #FIXME plot_annotation("Verity astrocytes with astrocyte markers")
p29_ver
ggsave("Verity_reactive_astrocyte_plot.png",plot=p29_ver,device='png', path=plotting_path_Ver,width = 20, height = 20, units = 'cm' )

#set the idents to 0.3
Verity_Prion_comp_astrocytes@active.ident<-(Verity_Prion_comp_astrocytes@meta.data$RNA_snn_res.0.3)
levels(Verity_Prion_comp_astrocytes)


DefaultAssay(Verity_Prion_comp_astrocytes)<-'RNA'

#by treatment, show the genes DE in Prion condition. LogFC set to a *1 increase in fold change
Verity_Prion_comp_astrocytes@active.ident<-as.factor(Verity_Prion_comp_astrocytes$treatment)
Verity_astrocyte_prion_markers<-FindMarkers(Verity_Prion_comp_astrocytes, ident.1 = 'ME7', ident.2 = 'NBH',min.pct = 0.10, verbose = TRUE, test.use = "DESeq2")

#Sort the list in descending log2FC and ascending pvalues
Verity_astrocyte_prion_markers<-Verity_astrocyte_prion_markers[order(-Verity_astrocyte_prion_markers$avg_log2FC), ]


#Use biomart info
Verity_astrocyte_prion_markers$gene<-rownames(Verity_astrocyte_prion_markers)

#check for duplicates
Verity_astrocyte_prion_markers[duplicated(Verity_astrocyte_prion_markers$gene),]

Verity_astrocyte_prion_markers<-Verity_astrocyte_prion_markers[!duplicated(Verity_astrocyte_prion_markers$gene),]

#make list of genes
gene_ids <- unique(Verity_astrocyte_prion_markers$gene)
  
#get relevant info from the mart
gene_info <- getBM(attributes = c("external_gene_name", "description","chromosome_name","entrezgene_id"), filters = "external_gene_name", values = gene_ids, mart = m)
  
#apply the info to the gene list
Verity_astrocyte_prion_markers <- merge(Verity_astrocyte_prion_markers, gene_info, by.x = "gene", by.y = "external_gene_name", all.x = TRUE)
  

#Save the file
write.csv(Verity_astrocyte_prion_markers, file = paste0(plotting_path,"Verity_astrocyte_prion_markers.csv"))

save(Verity_astrocyte_prion_markers, file="./Verity/Data/Verity_astrocyte_prion_markers.Rdata")

table(Verity_Prion_comp_astrocytes$treatment)
```

## Functional annotation

### Address the potential non-mapping gene entrezIDs, make a gene list
with fewer NAs

```{r}

#make list of gene names not mapping
not_mapped<-(Verity_astrocyte_prion_markers[is.na(Verity_astrocyte_prion_markers$entrezgene_id),]$gene)
not_mapped<-not_mapped[!duplicated(not_mapped)]

# Create an empty data frame to store the mapped ENTREZID values
remapped <- data.frame(ALIAS = character(0), ENTREZID = character(0), stringsAsFactors = FALSE)

# Loop through each gene in 'not_mapped'
for (gene in not_mapped) {
  # Attempt to perform the mapping using tryCatch
  mapped_genes <- tryCatch(
    {
      bitr(
        gene,
        fromType = "ALIAS",
        toType = "ENTREZID",
        OrgDb = org.Mm.eg.db,
        drop = FALSE
      )
    },
    error = function(e) {
      # If an error occurs, handle it by returning a data frame with NA values
      message(paste("Error mapping gene:", gene))
      return(data.frame(ALIAS = gene, ENTREZID = NA, stringsAsFactors = FALSE))
    }
  )

  # Append the temporary data frame to 'remapped'
  remapped <- rbind(remapped, mapped_genes)
  Sys.sleep(0.5)
}

#Additionally, get biomart conversions for those still without entrez ids 
biomart_remapped<-getBM(attributes = c("external_gene_name","entrezgene_id"), filters = "external_gene_name", values = remapped[is.na(remapped$ENTREZID),]$ALIAS, mart = m)

#add any more which have been mapped
biomart_remapped<-biomart_remapped[!is.na(biomart_remapped$entrezgene_id),]

#put everything recovered into a single df
bloated_mapped_df <- merge(remapped, biomart_remapped, by.x = "ALIAS", by.y = "external_gene_name", all.x = TRUE)
bloated_mapped_df$entrezgene_id <- as.character(bloated_mapped_df$entrezgene_id)

#make a single column for everything recovered, remove old columns
condensed_mapped <- bloated_mapped_df %>%
  mutate(entrezgene_id = coalesce(ENTREZID, entrezgene_id))
condensed_mapped$entrezgene_id<-as.integer(condensed_mapped$entrezgene_id)
#remove old columns
condensed_mapped$ENTREZID<-ENTREZID<- NULL


#The final list of newly gnerated entrezIDs, removed NAs
c<-condensed_mapped[!is.na(condensed_mapped$entrezgene_id),] 

#Add them back to the original list of markers, remove eztraneous columns
Verity_astrocyte_prion_markers <- merge(Verity_astrocyte_prion_markers, c, by.x = "gene", by.y = "ALIAS", all.x = TRUE)
Verity_astrocyte_prion_markers$entrezgene_id<-coalesce(Verity_astrocyte_prion_markers$entrezgene_id.x, Verity_astrocyte_prion_markers$entrezgene_id.y)
Verity_astrocyte_prion_markers<-subset(Verity_astrocyte_prion_markers, select = -c(entrezgene_id.x, entrezgene_id.y))
```

### Discard any genes which don't have an ENTREZID, generate list of
background DEGs

```{r}
#Discard any genes which don't have an ENTREZID
Verity_astrocyte_prion_markers_entrez<-Verity_astrocyte_prion_markers[!is.na(Verity_astrocyte_prion_markers$entrezgene_id),]
save(Verity_astrocyte_prion_markers_entrez, file="./Verity/Data/Verity_astrocyte_prion_markers_entrez.Rdata")

#create the genelist background for Verity Up
## feature 1: numeric vector
geneList_VP <- Verity_astrocyte_prion_markers_entrez$avg_log2FC
## feature 2: named vector of enembl IDs
names(geneList_VP) <-Verity_astrocyte_prion_markers_entrez$entrezgene_id
## feature 3: decreasing order
geneList_VP <- sort(geneList_VP, decreasing = TRUE) #object of type double
```



### Subset upregulated genes and perform ORA

```{r}
#Create upregulated significant genes
# Subset by Upregulated genes for GO enrichment
Verity_astrocyte_prion_markers_up<-subset(Verity_astrocyte_prion_markers_entrez, subset=avg_log2FC>(0.25)&p_val_adj<0.05) #Added p value filter here

#Create a list of just the gene symbols 
Verity_Prion_DEGs_up = (Verity_astrocyte_prion_markers_up$entrezgene_id)


# Run GO over-representation on upregulated entrezIds, biological process selected from GO terms
VerityPrion_up_GO_ORA = enrichGO(gene = Verity_Prion_DEGs_up, OrgDb = org.Mm.eg.db, readable = T, ont= "BP", pvalueCutoff = 0.05, qvalueCutoff = 0.10)

save(VerityPrion_up_GO_ORA, file="./Verity/Data/VerityPrion_up_GO_ORA.Rdata")

#FIXME could reduce terms by using simplify
termism_test<-pairwise_termsim(VerityPrion_up_GO_ORA)
simplify_ver_test<-clusterProfiler::simplify(termism_test)#, cutoff=0.7, by="p.adjust", select_fun=min)
emapplot(simplify_ver_test)
```


#### GO ORA visualisation and saving

```{r}
p30_ver <- barplot(VerityPrion_up_GO_ORA , showCategory=20,font.size = rel(0.60))+plot_annotation("Verity prion ORA enriched GO terms")
p31_ver <- dotplot(VerityPrion_up_GO_ORA, showCategory=10)+plot_annotation("Verity prion ORA enriched GO terms")
p32_ver <- goplot(VerityPrion_up_GO_ORA, showCategory = 20)+plot_annotation("Verity prion ORA enriched GO terms")
p33_ver<- cnetplot(VerityPrion_up_GO_ORA, foldChange=geneList_VP)+plot_annotation("Verity prion ORA enriched GO terms")
p34_ver<-heatplot(VerityPrion_up_GO_ORA, showCategory=20
                  #,foldChange=geneList_VP
                  )+plot_annotation("Verity prion ORA enriched GO terms")


ggsave("Verity_Prion_GOR_barplot.png",plot=p30_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GOR_dotplot.png",plot=p31_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GOR_goplot.png",plot=p32_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GOR_cnetplot.png",plot=p33_ver,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GOR_heatplot.png",plot=p34_ver,device='png', path=plotting_path_Ver,width = 60, height = 15, units = 'cm')

p30_ver
p31_ver
p32_ver
p33_ver
p34_ver

```

#### Do the same for downregulated, although this code functions, results are not meaningful as gene list membership is very small and results should not be trusted

```{r}
#Create downregulated significant genes for GO enrichment
Verity_astrocyte_prion_markers_down<-subset(Verity_astrocyte_prion_markers_entrez, subset=avg_log2FC<(-0.25)&p_val_adj<0.05) #Added p value filter here

#Create a list of just the gene symbols 
Verity_Prion_DEGs_down = (Verity_astrocyte_prion_markers_down$entrezgene_id)


# Run GO over-representation on upregulated entrezIds, biological process selected from GO terms
VerityPrion_down_GO_ORA = enrichGO(gene = Verity_Prion_DEGs_down, OrgDb = org.Mm.eg.db, readable = T, ont= "BP", pvalueCutoff = 0.05, qvalueCutoff = 0.10)


save(VerityPrion_down_GO_ORA, file="./Verity/Data/VerityPrion_down_GO_ORA.Rdata")
# #Uncomment to generate and save images, although these results are misleading.
# ggp_b<-barplot(VerityPrion_down_GO_ORA , showCategory=20)
# ggp2_b<-dotplot(VerityPrion_down_GO_ORA, showCategory=10)
# ggp3_b<-goplot(VerityPrion_down_GO_ORA, showCategory = 20)
# ggp4_b<-cnetplot(VerityPrion_down_GO_ORA, foldChange=geneList_VP)
# ggp5_b<-heatplot(VerityPrion_down_GO_ORA, showCategory=20)
# 
# ggp_b
# ggp2_b
# ggp3_b
# ggp4_b
# ggp5_b
# 
# ggsave("Verity_Prion_GOR_down_barplot.png",plot=ggp_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
# ggsave("Verity_Prion_GOR_down_dotplot.png",plot=ggp2_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
# ggsave("Verity_Prion_GOR_down_goplot.png",plot=ggp3_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
# ggsave("Verity_Prion_GOR_down_cnetplot.png",plot=ggp4_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
# ggsave("Verity_Prion_GOR_down_heatplot.png",plot=ggp5_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
# 
# V_P_topGO<-ggp2 + ggp2_b & plot_annotation(title="Verity Top and bottom most enriched GO terms",tag_levels = 'A')
# ggsave("V_P_topGO.png",plot=V_P_topGO,device='png', path=writing_dir,width = 40, height = 15, units = 'cm' )
# V_P_topGO
```


### GO GSEA

```{r}
#Perform GO GSEA
VP_GO_GSEA <- gseGO(geneList= geneList_VP, #This is the DEG list full list of DEGs
              OrgDb        = org.Mm.eg.db,
              ont          = "BP",
              minGSSize    = 100,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE,
              seed=42)

#convert back to Symbols for plotting
VP_GO_GSEA <- setReadable(VP_GO_GSEA, 'org.Mm.eg.db', 'ENTREZID')

save(VP_GO_GSEA, file="./Verity/Data/VP_GO_GSEA.Rdata")

View(data.frame(VP_GO_GSEA))
VP_GSEA_df<-data.frame(VP_GO_GSEA)
#search for the GSEA terms of a gene of interest
VP_GSEA_df[grep("Gfap", VP_GSEA_df$core_enrichment), ]$Description


write.csv(data.frame(VP_GO_GSEA), file = paste0(plotting_path,"VP_GO_GSEA.csv"))
```

#### GO GSEA visualisation

```{r}
ggp6<-dotplotGsea(data = VP_GO_GSEA,topn = 10,order.by = 'NES')#
#ggp6<-ggp6+ggtitle("Verity Prion comparison GO GSEA", subtitle="Top 10 enriched terms")

ggp7<-cnetplot(VP_GO_GSEA, colorEdge = TRUE,foldChange=geneList_VP, showCategory=5)+theme(plot.background = element_rect(fill = "white"))+plot_annotation(title = "Verity Prion comparison GO GSEA",  subtitle = "Gene-concept network plot")

ggp6$plot
ggp7

#calculate pairwise similarity of GSEA genesets
VP_GO_GSEA_interactions <- pairwise_termsim(VP_GO_GSEA)
#interaction between genesets
ggp7_1<-emapplot(VP_GO_GSEA_interactions, color ='NES',layout='nicely')+plot_annotation(title="Verity GO GSEA enrichment map", subtitle="Showing overlapping gene sets")

ggp7_1

ggsave("VP_GO_GSEA_dotplot.png",plot=ggp6$plot,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("VP_GO_GSEA_cnetplot.png",plot=ggp7,device='png', path=plotting_path_Ver,width = 50, height = 33, units = 'cm')
ggsave("VP_GO_GSEA_interactions.png",plot=ggp7_1,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
```

### Reactome ORA and GSEA

```{r}

#Reactome over-representation
VP_Reactome_ORA <- enrichPathway(gene=Verity_astrocyte_prion_markers_up$entrezgene_id, pvalueCutoff = 0.05, readable=TRUE, organism = "mouse")

#Reactome GSEA calculation
VP_Reactome_GSEA <- gsePathway(geneList_VP, 
                pvalueCutoff = 0.05,
                pAdjustMethod = "BH",
                organism = "mouse",
                verbose = FALSE,
                seed=42,
                maxGSSize = 100) ##FIXME trying a lower limit to remove broad terms

#convert back to symbols
VP_Reactome_GSEA <- setReadable(VP_Reactome_GSEA, 'org.Mm.eg.db', 'ENTREZID')
View(data.frame(VP_Reactome_GSEA))
write.csv(data.frame(VP_Reactome_GSEA), file = paste0(plotting_path,"VP_Reactome_GSEA.csv"))
```

#### Reactome GSEA visualisation

```{r}
#calculate pairwise similarity of GSEA genesets
VP_Reactome_GSEA_interactions <- pairwise_termsim(VP_Reactome_GSEA)

ggp8<-heatplot(VP_Reactome_GSEA, showCategory = 10, foldChange=geneList_VP)
ggp9<-cnetplot(VP_Reactome_GSEA, circular = TRUE, colorEdge = TRUE, foldChange=geneList_VP)
ggp10<-cnetplot(VP_Reactome_GSEA, colorEdge = TRUE, foldChange=geneList_VP, showCategory=10)
ggp11<-ridgeplot(VP_Reactome_GSEA,fill = 'NES', showCategory = 10)

#interaction between genesets
ggp12<-emapplot(VP_Reactome_GSEA_interactions, color ='NES',layout='nicely')

ggp8
ggp9
ggp10
ggp11
ggp12

ggsave("Verity_Prion_GSEA_heatplot_Reactome.png",plot=ggp8,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GSEA_cnetplot_circular_Reactome.png",plot=ggp9,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GSEA_cnetplot_Reactome.png",plot=ggp10,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GSEA_ridgeplot_Reactome.png",plot=ggp11,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Prion_GSEA_interactions_plot_Reactome.png",plot=ggp12,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
```

# Ageing, code works here, but plots were not generated as they weren't used in analysis.

```{r}
#active assay RNA
DefaultAssay(Verity_Ageing_comp)<-"RNA"

#DimReduc again
#https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/06_SC_SCT_normalization.md
# Normalize the counts
Verity_Ageing_comp <- NormalizeData(Verity_Ageing_comp)
# Identify the most variable genes with defaults
Verity_Ageing_comp <- FindVariableFeatures(Verity_Ageing_comp, verbose = T)
		     
# Scale the counts
Verity_Ageing_comp <- ScaleData(Verity_Ageing_comp)
# Perform PCA
Verity_Ageing_comp <- RunPCA(Verity_Ageing_comp)
#Inspect the variability accounted for by the PCS
#It appears that 23 PCs are responsible, so using 30 from now on
ElbowPlot(Verity_Ageing_comp, ndims=30)+ggtitle("Elbowplot of Verity data", subtitle = "NBH and ME7")
DimHeatmap(Verity_Ageing_comp, dims=1:9, cells=500, balanced=T)
#Finish dimensinoality reduction steps
Verity_Ageing_comp <- RunUMAP(Verity_Ageing_comp, reduction = "pca", dims = 1:30, seed.use=42)
Verity_Ageing_comp <- FindNeighbors(Verity_Ageing_comp, reduction = "pca", dims = 1:30)
Verity_Ageing_comp <- FindClusters(Verity_Ageing_comp, resolution = c(0.2,0.3,0.5,0.7,1))
#Save the object
save(Verity_Ageing_comp,file="./Verity/Data/Verity_Ageing_comp.Rdata")


# Visualise the clustering resolutions for cluster number

# Determine metrics to plot present in Ver_singlets_merged_clean@meta.data
metrics <-  c("nUMI", "nGene", "mitoRatio")


DimPlot(Verity_Ageing_comp , reduction = "umap", group.by = "treatment", raster=FALSE)+ ggtitle("Verity Hippocampal cells labelled by treatment")

#cyclechange resolution number in group.by to see difference in clusters
DimPlot(Verity_Ageing_comp, group.by = 'RNA_snn_res.0.2', reduction = 'umap', label = TRUE, repel = TRUE, raster=F)+ggtitle("Verity Hippocampal cells labelled by cluster numbers", subtitle = "Clustering resolution 0.2")

#adapted from https://github.com/hbctraining/scRNA-seq_online/blob/master/lessons/08_SC_clustering_quality_control.md#segregation-of-clusters-by-various-sources-of-uninteresting-variation

#Explore any clustering mishaps by lookin for localisation of any of these traits in a given cluster
FeaturePlot(Verity_Ageing_comp, 
            reduction = "umap", 
            features = metrics,
            pt.size = 0.4, 
            order = TRUE,
            #min.cutoff = 'q10', #This was a default in the online script, I don't see the benefit of colouring less of the range of values
            label = TRUE)


## Save the singlets with a cluster resolution?
 #set the idents 
Verity_Ageing_comp@active.ident<-Verity_Ageing_comp@meta.data$RNA_snn_res.0.2
levels(Verity_Ageing_comp)

rm(metrics)
```

```{r}
DefaultAssay(Verity_Ageing_comp)<-"RNA"

#Check for a gene of interest in the data's geneset
Gene_Names<-Verity_Ageing_comp@assays[["RNA"]]@counts@Dimnames[[1]]
filtered_gene_names<-Gene_Names[!grepl("^\\d", Gene_Names)]
filtered_gene_names[grep("^Slc1a2", filtered_gene_names)]

#try to identify astrocytes  with genes
astrocyte_genes<-c("Aldh1l1","Gfap","Slc1a2", "Slc1a3", "Glul", "Cd44")

FeaturePlot(Verity_Ageing_comp,features=astrocyte_genes, raster=F)#+ggtitle("Verity data, labelled with astrocyte gene markers from ABCAM", subtitle="Clustering resolution 0.3, 20 PCs used")

#As a reminder...
DimPlot(Verity_Ageing_comp, group.by = 'RNA_snn_res.0.2', reduction = 'umap', label = TRUE, repel = TRUE, raster=F)+ggtitle("Integrated Verity data, labelled by cluster numbers", subtitle="Clustering resolution 0.2")

#No compare to the plots of astrocyte genes
VlnPlot(Verity_Ageing_comp, features = astrocyte_genes, group.by = 'RNA_snn_res.0.2', pt.size = 0.0)

#Glul is maybe a good indicator?
```

## (SingleR annotation)

Could go here, could be followed by bubbleplot
<https://rdrr.io/github/milescsmith/seuratBubblePlot/man/bubbleplot.html>

Alternatively, load information from the spreadsheet at
Analysis/ABCAM_markers.xlsx then bubble plot

## Investigate astrocytes

```{r}
# Convert cluster names to character
Idents(Verity_Ageing_comp) <- Verity_Ageing_comp$RNA_snn_res.0.2

#Subset out astrocytes
Verity_Ageing_comp_astrocytes<-subset(Verity_Ageing_comp, idents=c('4','18'))
```

## DEG work

```{r}


# Run the standard workflow for visualization and clustering
Verity_Ageing_comp_astrocytes <- ScaleData(Verity_Ageing_comp_astrocytes, verbose = FALSE)
Verity_Ageing_comp_astrocytes <- RunPCA(Verity_Ageing_comp_astrocytes, npcs = 30, verbose = FALSE)
ElbowPlot(Verity_Ageing_comp_astrocytes, ndims=30)+ggtitle("Elbowplot of subset Verity astrocytes", subtitle = "RunPCA npcs=30")

#15 PCs sems ok, so use 20 from now

#20 PCs seem to cover the variability
Verity_Ageing_comp_astrocytes <- RunUMAP(Verity_Ageing_comp_astrocytes, reduction = "pca", dims = 1:20, seed.use=42)
Verity_Ageing_comp_astrocytes <- FindNeighbors(Verity_Ageing_comp_astrocytes, reduction = "pca", dims = 1:20)
Verity_Ageing_comp_astrocytes <- FindClusters(Verity_Ageing_comp_astrocytes, resolution = c(0.2,0.3,0.5,0.7,1))

#save for reloading
save(Verity_Ageing_comp_astrocytes, file="./Verity/Data/Verity_Ageing_comp_astrocytes.Rdata")

#cyclechange resolution number in group.by to see difference in clusters
DimPlot(Verity_Ageing_comp_astrocytes, reduction = 'umap',  raster=F, group.by='treatment')+ggtitle("Verity data's CX astrocytes", subtitle="Labelled by treatment")

#Show the clustering of subset astrocytes
DimPlot(Verity_Ageing_comp_astrocytes, group.by = 'RNA_snn_res.0.2', reduction = 'umap', repel = TRUE, raster=F)

#Show sample disctribution in clusters of astrocytes
DimPlot(Verity_Ageing_comp_astrocytes, group.by = 'sample', reduction = 'umap', repel = TRUE, raster=F)

#Show astrocyte markers in astrocyte subset
FeaturePlot(Verity_Ageing_comp_astrocytes,features=c('Gfap','Cd44'), raster=F)

#set the idents to 0.2
Verity_Ageing_comp_astrocytes@active.ident<-(Verity_Ageing_comp_astrocytes@meta.data$RNA_snn_res.0.2)
levels(Verity_Ageing_comp_astrocytes)

DefaultAssay(Verity_Ageing_comp_astrocytes)<-'RNA'

#by treatment, show the genes DE in Prion condition. LogFC set to a *1 increase in fold change
Verity_Ageing_comp_astrocytes@active.ident<-as.factor(Verity_Ageing_comp_astrocytes$treatment)
Verity_astrocyte_ageing_markers<-FindMarkers(Verity_Ageing_comp_astrocytes, ident.1 = 'AGEING', ident.2 = 'NBH',min.pct = 0.10, verbose = TRUE, test.use = "DESeq2")

#Sort the list in descending log2FC and ascending pvalues
Verity_astrocyte_ageing_markers<-Verity_astrocyte_ageing_markers[order(-Verity_astrocyte_ageing_markers$avg_log2FC), ]


#Use biomart info
Verity_astrocyte_ageing_markers$gene<-rownames(Verity_astrocyte_ageing_markers)

gene_ids <- unique(Verity_astrocyte_ageing_markers$gene)


#make list of genes
gene_ids <- unique(Verity_astrocyte_ageing_markers$gene)
  
#get relevant info from the mart
gene_info <- getBM(attributes = c("external_gene_name", "description","chromosome_name","entrezgene_id"), filters = "external_gene_name", values = gene_ids, mart = m)

  
#apply the info to the gene list
Verity_astrocyte_ageing_markers <- merge(Verity_astrocyte_ageing_markers, gene_info, by.x = "gene", by.y = "external_gene_name", all.x = TRUE)
  

#Save the file
write.csv(Verity_astrocyte_ageing_markers, file = paste0(plotting_path,"Verity_astrocyte_ageing_markers.csv"))

save(Verity_astrocyte_ageing_markers, file="./Verity/Data/Verity_astrocyte_ageing_markers.Rdata")

table(Verity_Ageing_comp_astrocytes$treatment)
```

## Functional annotation

Ageing is not giving many genes with logfc2 \> 0.25, so functional
annotation not working for upregulated genes. Could do down?


### Address the potential non-mapping gene entrezIDs, make a gene list
with fewer NAs

```{r}

#make list of gene names not mapping
not_mapped<-(Verity_astrocyte_ageing_markers[is.na(Verity_astrocyte_ageing_markers$entrezgene_id),]$gene)
not_mapped<-not_mapped[!duplicated(not_mapped)]

# Create an empty data frame to store the mapped ENTREZID values
remapped <- data.frame(ALIAS = character(0), ENTREZID = character(0), stringsAsFactors = FALSE)

# Loop through each gene in 'not_mapped'
for (gene in not_mapped) {
  # Attempt to perform the mapping using tryCatch
  mapped_genes <- tryCatch(
    {
      bitr(
        gene,
        fromType = "ALIAS",
        toType = "ENTREZID",
        OrgDb = org.Mm.eg.db,
        drop = FALSE
      )
    },
    error = function(e) {
      # If an error occurs, handle it by returning a data frame with NA values
      message(paste("Error mapping gene:", gene))
      return(data.frame(ALIAS = gene, ENTREZID = NA, stringsAsFactors = FALSE))
    }
  )

  # Append the temporary data frame to 'remapped'
  remapped <- rbind(remapped, mapped_genes)
  Sys.sleep(0.5)
}

#Additionally, get biomart conversions for those still without entrez ids 
biomart_remapped<-getBM(attributes = c("external_gene_name","entrezgene_id"), filters = "external_gene_name", values = remapped[is.na(remapped$ENTREZID),]$ALIAS, mart = m)

#add any more which have been mapped
biomart_remapped<-biomart_remapped[!is.na(biomart_remapped$entrezgene_id),]

#put everything recovered into a single df
bloated_mapped_df <- merge(remapped, biomart_remapped, by.x = "ALIAS", by.y = "external_gene_name", all.x = TRUE)
bloated_mapped_df$entrezgene_id <- as.character(bloated_mapped_df$entrezgene_id)

#make a single column for everything recovered, remove old columns
condensed_mapped <- bloated_mapped_df %>%
  mutate(entrezgene_id = coalesce(ENTREZID, entrezgene_id))
condensed_mapped$entrezgene_id<-as.integer(condensed_mapped$entrezgene_id)
#remove old columns
condensed_mapped$ENTREZID<-ENTREZID<- NULL


#The final list of newly gnerated entrezIDs, removed NAs
c<-condensed_mapped[!is.na(condensed_mapped$entrezgene_id),] 

#Add them back to the original list of markers, remove eztraneous columns
Verity_astrocyte_ageing_markers <- merge(Verity_astrocyte_ageing_markers, c, by.x = "gene", by.y = "ALIAS", all.x = TRUE)
Verity_astrocyte_ageing_markers$entrezgene_id<-coalesce(Verity_astrocyte_ageing_markers$entrezgene_id.x, Verity_astrocyte_ageing_markers$entrezgene_id.y)
Verity_astrocyte_ageing_markers<-subset(Verity_astrocyte_ageing_markers, select = -c(entrezgene_id.x, entrezgene_id.y))
```

### Discard any genes which don't have an ENTREZID, generate list of
background DEGs

```{r}
#Discard any genes which don't have an ENTREZID
Verity_astrocyte_ageing_markers_entrez<-Verity_astrocyte_ageing_markers[!is.na(Verity_astrocyte_ageing_markers$entrezgene_id),]


#create the genelist background for Verity Up
## feature 1: numeric vector
geneList_VA <- Verity_astrocyte_ageing_markers_entrez$avg_log2FC
## feature 2: named vector of enembl IDs
names(geneList_VA) <-Verity_astrocyte_ageing_markers_entrez$entrezgene_id
## feature 3: decreasing order
geneList_VA <- sort(geneList_VA, decreasing = TRUE) #object of type double
```

### Subset upregulated genes and perform ORA

```{r}
#Create upregulated significant genes
# Subset by Upregulated genes for GO enrichment
Verity_astrocyte_ageing_markers_up<-subset(Verity_astrocyte_ageing_markers_entrez, subset=avg_log2FC>(0.25)&p_val_adj<0.05) #Added p value filter here

#Create a list of just the gene symbols 
Verity_Ageing_DEGs_up = (Verity_astrocyte_ageing_markers_up$entrezgene_id)


# Run GO over-representation on upregulated entrezIds, biological process selected from GO terms
VerityAgeing_up_GO_ORA = enrichGO(gene = Verity_Ageing_DEGs_up, OrgDb = org.Mm.eg.db, readable = T, ont= "BP", pvalueCutoff = 0.05, qvalueCutoff = 0.10)
```

#### GO ORA visualisation and saving

```{r}
ggp <- barplot(VerityAgeing_up_GO_ORA , showCategory=20)
ggp2 <- dotplot(VerityAgeing_up_GO_ORA, showCategory=20)
ggp3 <- goplot(VerityAgeing_up_GO_ORA, showCategory = 20)
ggp4<- cnetplot(VerityAgeing_up_GO_ORA, foldChange=geneList_VA)
ggp5<-heatplot(VerityAgeing_up_GO_ORA, showCategory=20,foldChange=geneList_VA)


ggp
ggp2
ggp3
ggp4
ggp5

ggsave("Verity_Ageing_GOR_barplot.png",plot=ggp,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_dotplot.png",plot=ggp2,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_goplot.png",plot=ggp3,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_cnetplot.png",plot=ggp4,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_heatplot.png",plot=ggp5,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
```


#### Same for downregulation

```{r}
#Create downregulated significant genes for GO enrichment
Verity_astrocyte_ageing_markers_down<-subset(Verity_astrocyte_ageing_markers_entrez, subset=avg_log2FC<(-0.25)&p_val_adj<0.05) #Added p value filter here

#Create a list of just the gene symbols 
Verity_Ageing_DEGs_down = (Verity_astrocyte_ageing_markers_down$entrezgene_id)


# Run GO over-representation on upregulated entrezIds, biological process selected from GO terms
VerityAgeing_down_GO_ORA = enrichGO(gene = Verity_Ageing_DEGs_down, OrgDb = org.Mm.eg.db, readable = T, ont= "BP", pvalueCutoff = 0.05, qvalueCutoff = 0.10)

ggp_b<-barplot(VerityAgeing_down_GO_ORA , showCategory=20)
ggp2_b<-dotplot(VerityAgeing_down_GO_ORA, showCategory=20)
ggp3_b<-goplot(VerityAgeing_down_GO_ORA, showCategory = 20)
ggp4_b<-cnetplot(VerityAgeing_down_GO_ORA, foldChange=geneList_VP)
ggp5_b<-heatplot(VerityAgeing_down_GO_ORA, showCategory=20)

ggp_b
ggp2_b
ggp3_b
ggp4_b
ggp5_b

ggsave("Verity_Ageing_GOR_down_barplot.png",plot=ggp_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_down_dotplot.png",plot=ggp2_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_down_goplot.png",plot=ggp3_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_down_cnetplot.png",plot=ggp4_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
ggsave("Verity_Ageing_GOR_down_heatplot.png",plot=ggp5_b,device='png', path=plotting_path_Ver,width = 20, height = 15, units = 'cm')
```


### GO GSEA

```{r}
#Perform GO GSEA
VA_GO_GSEA <- gseGO(geneList= geneList_VA, #This is the DEG list full list of DEGs
              OrgDb        = org.Mm.eg.db,
              ont          = "BP",
              minGSSize    = 100,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE,
              seed=42)

#convert back to Symbols for plotting
VA_GO_GSEA <- setReadable(VA_GO_GSEA, 'org.Mm.eg.db', 'ENTREZID')

View(data.frame(VA_GO_GSEA))

write.csv(data.frame(VA_GO_GSEA), file = paste0(plotting_path,"VA_GO_GSEA.csv"))
```



#### GO GSEA visualisation

```{r}
ggp6<-dotplotGsea(data = VA_GO_GSEA,topn = 10,
            order.by = 'NES')

ggp7<-cnetplot(VA_GO_GSEA, colorEdge = TRUE, circular = TRUE,foldChange=geneList_VA, showCategory=5)


#calculate pairwise similarity of GSEA genesets
VA_GO_GSEA_interactions <- pairwise_termsim(VA_GO_GSEA)
#interaction between genesets
ggp7_1<-emapplot(VP_Reactome_GSEA_interactions, color ='NES',layout='nicely')

ggp6
ggp7
ggp7_1

png(paste0(plotting_path,"VA_GO_GSEA_dotplot_Reactome.png"), width = 1000,height = 750)
ggp6
dev.off()
png(paste0(plotting_path,"VA_GO_GSEA_cnetplot_Reactome.png"), width = 2000,height = 2000)
ggp7
dev.off()
png(paste0(plotting_path,"VA_GO_GSEA_interactions_Reactome.png"), width = 1500,height = 1500)
ggp7_1
dev.off()

```

### Reactome ORA and GSEA

```{r}

#Reactome over-representation
VA_Reactome_ORA <- enrichPathway(gene=Verity_astrocyte_ageing_markers_up$entrezgene_id, pvalueCutoff = 0.05, readable=TRUE, organism = "mouse")

#Reactome GSEA calculation
VA_Reactome_GSEA <- gsePathway(geneList_VA, 
                pvalueCutoff = 0.05,
                pAdjustMethod = "BH",
                organism = "mouse",
                verbose = FALSE,
                seed=42,
                maxGSSize = 100) ##FIXME trying a lower limit to remove broad terms

#convert back to symbols
VA_Reactome_GSEA <- setReadable(VA_Reactome_GSEA, 'org.Mm.eg.db', 'ENTREZID')
View(data.frame(VA_Reactome_GSEA))
write.csv(data.frame(VA_Reactome_GSEA), file = paste0(plotting_path,"VA_Reactome_GSEA.csv"))
```

#### Reactome GSEA visualisation

```{r}
#calculate pairwise similarity of GSEA genesets
VA_Reactome_GSEA_interactions <- pairwise_termsim(VA_Reactome_GSEA)

ggp8<-heatplot(VA_Reactome_GSEA, showCategory = 10, foldChange=geneList_VA)
ggp9<-cnetplot(VA_Reactome_GSEA, circular = TRUE, colorEdge = TRUE, foldChange=geneList_VA)
ggp10<-cnetplot(VA_Reactome_GSEA, colorEdge = TRUE, foldChange=geneList_VA, showCategory=10)
ggp11<-ridgeplot(VA_Reactome_GSEA,fill = 'NES', showCategory = 10)

#interaction between genesets
ggp12<-emapplot(VA_Reactome_GSEA_interactions, color ='NES',layout='nicely')

ggp8
ggp9
ggp10
ggp11
ggp12

png(paste0(plotting_path,"Verity_Ageing_GSEA_heatplot.png"), width = 1000,height = 750)
ggp8
dev.off()
png(paste0(plotting_path,"Verity_Ageing_GSEA_cnetplot_circular.png"), width =2000,height = 2000)
ggp9
dev.off()
png(paste0(plotting_path,"Verity_Ageing_GSEA_cnetplot.png"), width =2000,height = 2000)
ggp10
dev.off()
png(paste0(plotting_path,"Verity_Ageing_GSEA_ridgeplot.png"), width = 1000,height = 750)
ggp11
dev.off()
png(paste0(plotting_path,"Verity_Ageing_GSEA_interactions_plot.png"), width = 1500,height = 1500)
ggp12
dev.off()

```
